Adding reCAPTCHA V3 into Payment Page: - move verify captcha after submit clicked - adding threshold on properties as Score validation - saving captcha_score and captcha_token into table transaction - show captcha_score and captcha_token on transaction detail component
WARNING: there's database schema changes, need to alter the table or remove h2 database
ALTER TABLE transaction ADD COLUMN captcha_token varchar(1000);
ALTER TABLE transaction ADD COLUMN captcha_score float4;

+++ src/main/java/men/doku/donation/service/dto/RecaptchaVerifyResponseDTO.java
@@ -16,7 +16,7 @@ public class RecaptchaVerifyResponseDTO {
     @JsonProperty("error-codes")
     String[] errorCodes;
 
-    Long score;
+    Float score;
 
     String action;
 
@@ -24,7 +24,7 @@ public class RecaptchaVerifyResponseDTO {
     }
 
     public RecaptchaVerifyResponseDTO(Boolean success, String challengeTs, String hostname, String[] errorCodes,
-            Long score, String action) {
+            Float score, String action) {
         this.success = success;
         this.challengeTs = challengeTs;
         this.hostname = hostname;
@@ -65,11 +65,11 @@ public class RecaptchaVerifyResponseDTO {
         this.errorCodes = errorCodes;
     }
 
-    public Long getScore() {
+    public Float getScore() {
         return score;
     }
 
-    public void setScore(Long score) {
+    public void setScore(Float score) {
         this.score = score;
     }
